requirements.txt
gradio
python-dotenv
requests
google-adk
google-generativeai
pydantic
pillow
matplotlib
fpdf==1.7.2